+++ lib/src/video_player_view.dart
@@ -132,14 +132,21 @@ class VideoPlayerView extends StatelessWidget {
           onPlatformViewCreated: _onPlatformViewCreated,
           creationParamsCodec: const StandardMessageCodec(),
         );
-      case TargetPlatform.fuchsia:
-        return Text('$defaultTargetPlatform is not yet supported by the web_view plugin');
-      case TargetPlatform.windows:
-        return Text('$defaultTargetPlatform is not yet supported by the web_view plugin');
-      case TargetPlatform.linux:
-        return Text('$defaultTargetPlatform is not yet supported by the web_view plugin');
       case TargetPlatform.macOS:
-        return Text('$defaultTargetPlatform is not yet supported by the web_view plugin');
+        return AppKitView(
+          viewType: _viewType,
+          layoutDirection: TextDirection.ltr,
+          hitTestBehavior: PlatformViewHitTestBehavior.transparent,
+          creationParams: <String, dynamic>{'url': url, 'resizeMode': resizeMode.value},
+          onPlatformViewCreated: _onPlatformViewCreated,
+          creationParamsCodec: const StandardMessageCodec(),
+        );
+      case TargetPlatform.fuchsia:
+        return Text('$defaultTargetPlatform is not yet supported by the video_player plugin');
+      case TargetPlatform.windows:
+        return Text('$defaultTargetPlatform is not yet supported by the video_player plugin');
+      case TargetPlatform.linux:
+        return Text('$defaultTargetPlatform is not yet supported by the video_player plugin');
     }
   }